Move version resolution focused tests
* Moved to a dedicated versions package
